refactor: Move api calls to separate directory
