+++ src/api/post-pacts.ts
@@ -1,4 +1,4 @@
-import { ServiceContractsForm } from './generate-contracts-form';
+import { ServiceContractsForm } from '../utils/generate-contracts-form';
 import axios from 'axios';
 
 export function postPacts(url: string, pacts: ServiceContractsForm) {
