+++ test/post-pacts.test.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { serviceContractsFormMockFactory } from './mocks/serviceContractsForm.mock';
-import { postPacts } from '../src/utils/post-pacts';
+import { postPacts } from '../src/api/post-pacts';
 
 jest.mock('axios');
